# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "paths": {
    "/api/v1/auth/login": {
      "post": {
        "tags": [
          "authentication"
        ],
        "summary": "Bearer Login",
        "operationId": "bearer_login_api_v1_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/loginIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/refresh": {
      "post": {
        "tags": [
          "authentication"
        ],
        "summary": "Refresh Session",
        "operationId": "refresh_session_api_v1_auth_refresh_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "tags": [
          "authentication"
        ],
        "summary": "Logout",
        "operationId": "logout_api_v1_auth_logout_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LogoutRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/login": {
      "post": {
        "summary": "Login",
        "operationId": "login_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/loginIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/loginResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Empresa": {
        "properties": {
          "uuid": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Uuid"
          },
          "cnpj": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cnpj"
          },
          "razaoSocial": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Razaosocial"
          },
          "nomeFantasia": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nomefantasia"
          },
          "municipio": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Municipio"
          },
          "uf": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Uf"
          },
          "cnae": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cnae"
          },
          "ativo": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ativo"
          },
          "bloqueado": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bloqueado"
          },
          "motivoBloqueio": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Motivobloqueio"
          },
          "plano": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Plano"
          },
          "statusAssinatura": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Statusassinatura"
          },
          "dataInicioAssinatura": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Datainicioassinatura"
          },
          "dataFimAssinatura": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Datafimassinatura"
          },
          "origemAssinatura": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Origemassinatura"
          },
          "dataCadastro": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Datacadastro"
          },
          "dataAtualizacao": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dataatualizacao"
          },
          "sincronizado": {
            "type": "boolean",
            "title": "Sincronizado"
          }
        },
        "type": "object",
        "required": [
          "sincronizado"
        ],
        "title": "Empresa"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LogoutRequest": {
        "properties": {
          "refreshToken": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 512,
                "minLength": 43
              },
              {
                "type": "null"
              }
            ],
            "title": "Refreshtoken"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "LogoutRequest"
      },
...
      "RefreshRequest": {
        "properties": {
          "refreshToken": {
            "type": "string",
            "maxLength": 512,
            "minLength": 43,
            "title": "Refreshtoken"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "refreshToken"
        ],
        "title": "RefreshRequest"
      },
...
      "Usuario": {
        "properties": {
          "uuid": {
            "type": "string",
            "title": "Uuid"
          },
          "email": {
            "type": "string",
            "title": "Email"
          },
          "senhaHash": {
            "type": "string",
            "title": "Senhahash"
          },
          "nome": {
            "type": "string",
            "title": "Nome"
          },
          "empresaUuid": {
            "type": "string",
            "title": "Empresauuid"
          },
          "ativo": {
            "type": "boolean",
            "title": "Ativo"
          },
          "dataCadastro": {
            "type": "integer",
            "title": "Datacadastro"
          },
          "ultimoLogin": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ultimologin"
          }
        },
        "type": "object",
        "required": [
          "uuid",
          "email",
          "senhaHash",
          "nome",
          "empresaUuid",
          "ativo",
          "dataCadastro"
        ],
        "title": "Usuario"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "loginIn": {
        "properties": {
          "login": {
            "type": "string",
            "title": "Login"
          },
          "senha": {
            "type": "string",
            "title": "Senha"
          }
        },
        "type": "object",
        "required": [
          "login",
          "senha"
        ],
        "title": "loginIn"
      },
      "loginResponse": {
        "properties": {
          "sucesso": {
            "type": "string",
            "title": "Sucesso"
          },
          "mensagem": {
            "type": "string",
            "title": "Mensagem"
          },
          "usuario": {
            "$ref": "#/components/schemas/Usuario"
          },
          "empresa": {
            "$ref": "#/components/schemas/Empresa"
          }
        },
        "type": "object",
        "required": [
          "sucesso",
          "mensagem",
          "usuario",
          "empresa"
        ],
        "title": "loginResponse"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}